requirements.txt
redis
datetime
requests
python-dotenv